Scope self locally to prevent global contamination
It turned out the getData() of one of my other modules wasn't called anymore because we both use self to get the address of the getData() callback function. It turns out that, without ```var``` nodejs uses a globally scoped self and since your module was loaded later, it allways "won" the callBack function. Using ```var``` in both your and my module keeps the callbacks sane again.